Add shared `useBeforeUnloadPromptEffect`.
Use it in edit flow.

(cherry picked from commit 60c1f0e642b558e85d8ad15fc8fd470df61fffca)
