+++ js/src/setup-ads/setup-ads-form-content.js
@@ -1,11 +1,7 @@
 /**
- * External dependencies
- */
-import { useEffect } from '@wordpress/element';
-
-/**
  * Internal dependencies
  */
+import useBeforeUnloadPromptEffect from '.~/hooks/useBeforeUnloadPromptEffect';
 import AdsStepper from './ads-stepper';
 import isFormDirty from './is-form-dirty';
 import SetupAdsTopBar from './top-bar';
@@ -14,23 +10,7 @@ const SetupAdsFormContent = ( props ) => {
 	const { formProps } = props;
 	const shouldPreventClose = isFormDirty( formProps );
 
-	useEffect( () => {
-		const eventListener = ( e ) => {
-			// If you prevent default behavior in Mozilla Firefox prompt will always be shown.
-			e.preventDefault();
-
-			// Chrome requires returnValue to be set.
-			e.returnValue = '';
-		};
-
-		if ( shouldPreventClose ) {
-			window.addEventListener( 'beforeunload', eventListener );
-		}
-
-		return () => {
-			window.removeEventListener( 'beforeunload', eventListener );
-		};
-	}, [ shouldPreventClose ] );
+	useBeforeUnloadPromptEffect( shouldPreventClose );
 
 	return (
 		<>
